Queue File
The Google App Engine uses something called “Task Queues” that are
configured by this file.

When the application is running it checks the Yahoo weather API every 5
minutes to see if it is raining.

If it IS NOT raining then it will log a message and reload the
application into the task queue.

If it IS raining then it will send a tweet and PURGE the task queue so
that it doesn’t keep tweeting every 5 minutes.

The Cron file controls when the application starts over each day.